✅ migrate payment_request e2e to ts + fixtures
